Refactor: 파일, 폴더구조 변경
