+++ app/(route)/main/_components/Banner/index.tsx
