+++ app/(route)/main/_components/HotActivityLists/index.tsx
