+++ app/(route)/main/_components/SearchBox/index.tsx
